+++ home.admin/config.scripts/lnd.credentials.sh
@@ -26,10 +26,10 @@ source <(/home/admin/config.scripts/network.aliases.sh getvars lnd ${CHAIN})
 function copy_mac_set_perms() {
   local file_name=${1}  # the file name (e.g. admin.macaroon)
   local group_name=${2} # the unix group name (e.g. lndadmin)
-  local n=${3:-glcoin} # the network (e.g. glcoin) defaults to glcoin
+  local n=${3:-mainnet} # the network subdirectory under chain/glcoin (e.g. mainnet, testnet)
   local c=${4:-main}    # the chain (e.g. main, test, sim, reg) defaults to main (for mainnet)
-  sudo /bin/chown --silent admin:"${group_name}" /mnt/hdd/app-data/lnd/data/chain/bitcoin/"${n}"/"${file_name}"
-  sudo /bin/chmod --silent 750 /mnt/hdd/app-data/lnd/data/chain/bitcoin/"${n}"/"${file_name}"
+  sudo /bin/chown --silent admin:"${group_name}" /mnt/hdd/app-data/lnd/data/chain/glcoin/"${n}"/"${file_name}"
+  sudo /bin/chmod --silent 750 /mnt/hdd/app-data/lnd/data/chain/glcoin/"${n}"/"${file_name}"
 }
 
 function check_macaroons() {
@@ -38,9 +38,9 @@ missing=0
 for macaroon in $macaroons
 do
   local file_name=${macaroon}
-  local n=${1:-glcoin} # the network (e.g. glcoin) defaults to glcoin
+  local n=${1:-mainnet} # the network subdirectory under chain/glcoin (e.g. mainnet, testnet)
   local c=${2:-main}    # the chain (e.g. main, test, sim, reg) defaults to main (for mainnet)
-  if [ ! -f /mnt/hdd/app-data/lnd/data/chain/bitcoin/"${n}"/"${macaroon}" ]; then
+  if [ ! -f /mnt/hdd/app-data/lnd/data/chain/glcoin/"${n}"/"${macaroon}" ]; then
     missing=$((missing + 1))
     echo "# ${macaroon} is missing ($missing)"
   else
@@ -84,11 +84,9 @@ if [ "$1" = "reset" ]; then
     echo "## Resetting Macaroons"
     echo "# all your macaroons get deleted and recreated"
     cd || exit
-    # shellcheck disable=SC2154 # gets the ${network} from the raspiblesk.conf
-    sudo find /mnt/hdd/app-data/lnd/data/chain/bitcoin/"${network}"/ -iname '*.macaroon' -delete
-    sudo find /mnt/hdd/app-data/lnd/data/chain/bitcoin/"${network}"/ -iname '*.macaroon' -delete
+    sudo find /mnt/hdd/app-data/lnd/data/chain/glcoin/"${CHAIN}"/ -iname '*.macaroon' -delete
     if [ "${keepOldMacaroons}" != "1" ]; then
-      sudo rm /mnt/hdd/app-data/lnd/data/chain/bitcoin/"${network}"/macaroons.db
+      sudo rm /mnt/hdd/app-data/lnd/data/chain/glcoin/"${CHAIN}"/macaroons.db
     fi
 
     echo "# delete also lit macaroons if present"
@@ -114,9 +112,9 @@ if [ "$1" = "reset" ]; then
 
   if [ ${resetMacaroons} -eq 1 ]; then
     echo "# copy new macaroons to central app-data directory and ensure unix ownerships and permissions"
-    copy_mac_set_perms admin.macaroon lndadmin "${network}" "${chain}"
-    copy_mac_set_perms invoice.macaroon lndinvoice "${network}" "${chain}"
-    copy_mac_set_perms readonly.macaroon lndreadonly "${network}" "${chain}"
+    copy_mac_set_perms admin.macaroon lndadmin "${CHAIN}" "${chain}"
+    copy_mac_set_perms invoice.macaroon lndinvoice "${CHAIN}" "${chain}"
+    copy_mac_set_perms readonly.macaroon lndreadonly "${CHAIN}" "${chain}"
     echo "# OK DONE"
   fi
 
@@ -133,7 +131,7 @@ elif [ "$1" = "sync" ]; then
   echo "###### SYNCING MACAROONS, RPC Password AND TLS Certificate ######"
 
   echo "# make sure LND app-data directories exist"
-  sudo /bin/mkdir --mode 0755 --parents /mnt/hdd/app-data/lnd/data/chain/bitcoin/"${network}"/
+  sudo /bin/mkdir --mode 0755 --parents /mnt/hdd/app-data/lnd/data/chain/glcoin/"${CHAIN}"/
 
   echo `# make sure all user groups exist for default macaroons`
   sudo /usr/sbin/groupadd --force --gid 9700 lndadmin
@@ -153,14 +151,14 @@ elif [ "$1" = "sync" ]; then
   fi
 
   echo "# copy macaroons to central app-data directory and ensure unix ownerships and permissions"
-  copy_mac_set_perms admin.macaroon lndadmin "${network}" "${chain}"
-  copy_mac_set_perms invoice.macaroon lndinvoice "${network}" "${chain}"
-  copy_mac_set_perms readonly.macaroon lndreadonly "${network}" "${chain}"
-  copy_mac_set_perms invoices.macaroon lndinvoices "${network}" "${chain}"
-  copy_mac_set_perms chainnotifier.macaroon lndchainnotifier "${network}" "${chain}"
-  copy_mac_set_perms signer.macaroon lndsigner "${network}" "${chain}"
-  copy_mac_set_perms walletkit.macaroon lndwalletkit "${network}" "${chain}"
-  copy_mac_set_perms router.macaroon lndrouter "${network}" "${chain}"
+  copy_mac_set_perms admin.macaroon lndadmin "${CHAIN}" "${chain}"
+  copy_mac_set_perms invoice.macaroon lndinvoice "${CHAIN}" "${chain}"
+  copy_mac_set_perms readonly.macaroon lndreadonly "${CHAIN}" "${chain}"
+  copy_mac_set_perms invoices.macaroon lndinvoices "${CHAIN}" "${chain}"
+  copy_mac_set_perms chainnotifier.macaroon lndchainnotifier "${CHAIN}" "${chain}"
+  copy_mac_set_perms signer.macaroon lndsigner "${CHAIN}" "${chain}"
+  copy_mac_set_perms walletkit.macaroon lndwalletkit "${CHAIN}" "${chain}"
+  copy_mac_set_perms router.macaroon lndrouter "${CHAIN}" "${chain}"
 
   sudo usermod -aG lndadmin admin
   sudo usermod -aG glcoin admin
@@ -170,9 +168,9 @@ elif [ "$1" = "sync" ]; then
 # Check Macaroons and fix missing
 ###########################
 elif [ "$1" = "check" ]; then
-  check_macaroons "${network}" "${chain}"
+  check_macaroons "${CHAIN}" "${chain}"
   if [ $missing -gt 0 ]; then
-    /home/admin/config.scrips/lnd.credentials.sh reset "${CHAIN}" keepold
+    /home/admin/config.scripts/lnd.credentials.sh reset "${CHAIN}" keepold
   fi
 
 ###########################